Only set the TK iconbitmap on Windows
Unfortunately using an .ico file here doesn't play well on Linux, and the script was failing on this line for me.
On Linux it should be an .xbm image, but I was not able to successfully convert the current .ico file to a working .xbm - so for the moment, I've just skipped this setting for non-Windows OSs.

Refer to this Stack Overflow answer: https://stackoverflow.com/a/16161574